Implemented images for the cards, they are given in the backbone Card model.
To obtain the source information, I needed the Cards from the database,
so the album/home view also renders the cards in the router call.

A new global object has been creaded for this end, window.db
and is an object containing in the key "cards" a Backbone collection
with the card models. This is done to have a place to hold information
that will not change, like the cards in this case.

Now users, upon initialize, use the db, to complete their own models,
so for each user, when specifying their album, we only send the card_id
and the count, the rest of the information is extacted form window.db