Now the notifications in backbone are displayed using 2 templates, trade and recevied. The trade template is used when rendering a notification with sender_id and and the raceived template is used if the sender_id is not set.
Each template now listens to the accept or ok button. In the case of
acceptance, a post with the information of the notification is sent to
the server.

Also, to show the user portraits in the notifications, I included the
users field into the window.db objects. It's a Users collection and it's
filled with the friends of the user and the user himself. The idea is to
have a place where to store user information that should not change over
time.