Ruby:
Now the challenge_id attribute is changable in the notification model,
allowing for challenge completions when creating a new challenge-complete
notification.

I also included some validators in the notification creation,
 validate_trade and validate_challenge,  but only validate_trade is
implemented, validate_challenge is mearly a method that returns true

I also implemented the success challenge complete case in backbone, so
if a challenge is successfully completed, then the cards are updated.